feat: reorganize client sub module and all imports
