+++ genai_impact/__init__.py
@@ -1,3 +1,3 @@
-from genai_impact.client_wrapper import OpenAI
+from genai_impact.client import OpenAI
 
 __all__ = ["OpenAI"]
