+++ genai_impact/__main__.py
@@ -1,4 +1,4 @@
-from genai_impact.client_wrapper import OpenAI
+from genai_impact.client import OpenAI
 
 client = OpenAI()
