+++ genai_impact/client/__init__.py
@@ -1,0 +1,5 @@
+from .openai_wrapper import OpenAI
+
+__all__ = [
+    "OpenAI"
+]
